requirements.txt
Django>=5.1.1,<5.2
djangorestframework>=3.15.2,<3.16
psycopg2>=2.9.9,<3
drf-spectacular>=0.27.2,<0.30
Pillow>=10.4.0,<10.5.0
uWSGI>=2.0.26,<2.1.0